Add support to run gemini and chatterbot at the same time
